backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-MongoEngine==0.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
pymongo==4.6.0
mongoengine==0.27.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Cors==4.0.0
WTForms==3.0.1
email-validator==2.1.0
requests==2.31.0
PyJWT==2.8.0
pyzbar==0.1.9
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pytesseract>=0.3.10
fuzzywuzzy[speedup]>=0.18.0
python-Levenshtein>=0.21.0
pdfplumber>=0.9.0
python-dotenv>=1.0.1
reportlab>=4.1.0
pandas>=2.0.0
cloudinary>=1.29.0
